+++ packages/utils/src/sys/super/index.ts
@@ -1,3 +1,4 @@
 export * from './core'
+export * from './copy'
 export * from './os'
 export * from './env-paths'