+++ packages/_config/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Minor Changes
 
-- add utils package, change yargs to clippium etc
+- add utils package, change yargs to clippium, change LICENSE to MIT, add support for contentDirs etc
 
 ## 0.1.17